cast RAND_MAX to unsigned long for gcc 4.4

+++ Random/src/RandEngine.cc
@@ -1,4 +1,4 @@
-// $Id: RandEngine.cc,v 1.4.4.5.2.1 2008/11/13 21:35:23 garren Exp $
+// $Id: RandEngine.cc,v 1.4.4.5.2.2 2010/03/08 20:18:19 garren Exp $
 // -*- C++ -*-
 //
 // -----------------------------------------------------------------------
@@ -62,6 +62,8 @@ MTwistEngine * fakeFlat = new MTwistEngine;
 RandFlat rflat (fakeFlat, 0, RAND_MAX+1);   
 int rand() { return (int)rflat(); }         
 #endif                                      
+
+
 
 static const int MarkerLen = 64; // Enough room to hold a begin or end marker. 
 
@@ -254,7 +256,7 @@ void RandEngine::showStatus() const
   {                                                         
     // Here, we know that integer arithmetic is 64 bits.    
     if ( !prepared ) {                                      
-      iS = RAND_MAX + 1;                     
+      iS = (unsigned long)RAND_MAX + 1;                     
       iK = 1;                                
 //    int StoK = S;                          
       int StoK = iS;          
@@ -290,7 +292,7 @@ void RandEngine::showStatus() const
     // of precision, but we have no idea how many randoms we will need to 
     // generate 32 bits.                                                  
     if ( !prepared ) {                                                    
-      fS = RAND_MAX + 1;                                                  
+      fS = (unsigned long)RAND_MAX + 1;                                                  
       double twoTo32 = ldexp(1.0,32);                                     
       double StoK = fS;                                                   
       for ( iK = 1; StoK < twoTo32; StoK *= fS, iK++ ) { }                
@@ -317,7 +319,7 @@ void RandEngine::showStatus() const
     // Here, we know that 16 random bits are available from each of       
     // two random numbers.                                                
     if ( !prepared ) {                                                    
-      iS = RAND_MAX + 1;                                                  
+      iS = (unsigned long)RAND_MAX + 1;                                                  
       int SshiftN = iS;                                                   
       for (iN = 0; SshiftN > 1; SshiftN >>= 1, iN++) { }                  
       iN -= 17;
